feat: Prepare project for CMS (Sanity) integration using service & repository patterns
- Moved core data models & interfaces into a dedicated module structure
- Defined asynchronous Repository Interfaces for Breweries, Trails, and Guides (`lib/repositories/interfaces.ts`)
- Created local Mock Repository implementations (`lib/repositories/mock/index.ts`) backed by existing mock data
- Implemented content service layer (`lib/services/content.service.ts`) for unified dynamic fetching
- Refactored all data-dependent pages to fetch data asynchronously via `contentService`
- Extracted interactive/search logic to custom Client components to allow dynamic pages to be Next.js Server Components, optimizing SEO and performance
- Cleaned up page styles and verified production builds and frontend render successfully without direct imports of mock-data.ts in UI pages.
